+++ examples/resources/cvp_studio_inputs/import.sh
@@ -1,2 +1,2 @@
-# cvp_studio_inputs is imported by its composite id: {workspace_id}:{studio_id}:{path_joined}
-terraform import cvp_studio_inputs.rack3_bgp '8f3c0b2a-...:1dd135ac-...:tenants/[name=default]/vrfs/[name=RED]/lbBgp'
+# cvp_studio_inputs is imported by its composite id: {workspace_id}/{studio_id}/{path...}
+terraform import cvp_studio_inputs.rack3_bgp '8f3c0b2a-1d4e-4a9b-9c11-2bf1dbc3fa86/studio-l3ls/tenants/[name=default]/vrfs/[name=RED]/lbBgp'